feat: implementando o exception handler, o mediatr e os controllers

+++ src/Domain/Entities/CestaRecomendacao.cs
@@ -1,10 +1,8 @@
-using System.Numerics;
-
 namespace CompraProgamada.Domain.Entities;
 
 public class CestaRecomendacao
 {
-    public BigInteger Id { get; set; }
+    public long Id { get; set; }
     public string Nome { get; set; } = string.Empty;
     public bool Ativa { get; set; }
     public DateTime DataCriacao { get; set; }
