+++ src/Domain/Entities/Custodia.cs
@@ -1,11 +1,9 @@
-using System.Numerics;
-
 namespace CompraProgamada.Domain.Entities;
 
 public class Custodia
 {
-    public BigInteger Id { get; set; }
-    public BigInteger ContaGraficaId { get; set; }
+    public long Id { get; set; }
+    public long ContaGraficaId { get; set; }
     public string Ticker { get; set; } = string.Empty;
     public int Quantidade { get; set; }
     public decimal PrecoMedio { get; set; }
