+++ src/Domain/Entities/EventosIR.cs
@@ -1,12 +1,11 @@
-using System.Numerics;
 using CompraProgamada.Domain.Enums;
 
 namespace CompraProgamada.Domain.Entities;
 
 public class EventosIR
 {
-    public BigInteger Id { get; set; }
-    public BigInteger ClienteId { get; set; }
+    public long Id { get; set; }
+    public long ClienteId { get; set; }
     public TipoImpostoRenda TipoImposto { get; set; }
     public decimal ValorBase { get; set; }
     public decimal ValorImposto { get; set; }
